Refactored main class added class for MainWindow
